WebSocketApp: Allow callbacks to be declared as overridden methods
Previously, declaring a subclass of `WebSocketApp` would not allow callbacks,
such as `on_open` to be declared as overridden methods; this is because the
callback functions are accepted as arguments to the constructor and are
unconditonally assigned there. (The subclass could simply avoid calling its
superclass constructor, but that doesn't seem like a good approach.)

This commit therefore declares a stub method for each callback and makes the
assignments in the constructor conditional, thereby allowing subclasses to
override the callback methods directly.